requirements.txt
numpy
pygetwindow
sounddevice
wavio
openai-whisper
imageio-ffmpeg
